Make GCE its own section (with DNS) and other mainly empty docs
This will be populated over the coming days.
Docs once refined should probably slowly move their way from the
deploy repo to here?
Finding that balance will be hard?

I did want to move this docs repo to the deploy repo.
The downside of that would be in the future if deployments are
automated, then having the docs in there would be silly..